export data conversion functionality
