+++ packages/ricos-content/web/src/migrateSchema/index.ts
@@ -1,2 +1,7 @@
 export { fromDraft, ensureRicosContent } from './fromDraft/fromDraft';
+export { convertBlockDataToDraft } from './fromDraft/convertRicosPluginData';
 export { toDraft, ensureDraftContent } from './toDraft/toDraft';
+export {
+  convertNodeDataToDraft,
+  convertDecorationDataToDraft,
+} from './toDraft/convertDraftPluginData';
